Deserialize non-byte strings into raw buffers
Allow deserialization of non-byte string values into raw byte buffers.
In cases where a value is a non-byte string, a byte buffer can be used
to capture the raw encoded value. For instance, assuming a dictionary
with an `info` key which has a dictionary value:

```
struct Metainfo {
    info: ByteBuf,
}
```

could be used to capture the raw bytes of the encoded `info` dictionary
value.

For untrusted input, the value should be verified as having the correct
type (e.g. a dictionary) instead of a byte string which contains the raw
encoded value.